Move gitmetadata to internal
